requirements.txt
alabaster==0.7.4
Babel==1.3
cffi==0.9.2
cryptography==0.8.2
diff-match-patch==20121119
Django==1.7.4
django-cors-headers==1.0.0
django-debug-toolbar==1.3.0
django-devserver==0.8.0
django-extensions==1.5.0
django-import-export==0.2.6
django-model-utils==2.2
django-money==0.5.0
django-pgjson==0.2.3
django-photologue==3.1.1
django-registration-redux==1.1
django-sortedm2m==0.9.4
django-taggit==0.12.2
django-widget-tweaks==1.3
djangorestframework==3.1.1
djrill==1.3.0
docutils==0.12
drf-extensions==0.2.7
ecdsa==0.13
enum34==1.0.4
Fabric==1.10.1
gnureadline==6.3.3
guppy==0.1.10
ipython==3.1.0
Jinja2==2.7.3
jsonfield==1.0.2
line-profiler==1.0
MarkupSafe==0.23
ndg-httpsclient==0.4.0
paramiko==1.15.2
Pillow==2.7.0
psycopg2==2.6
py-moneyed==0.5.0
pyasn1==0.1.7
pycparser==2.12
pycrypto==2.6.1
Pygments==2.0.2
pyOpenSSL==0.15.1
pytz==2014.10
reportlab==3.1.44
requests==2.7.0
shortuuid==0.4.2
six==1.9.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.8
sqlparse==0.1.15
stripe==1.22.2
tablib==0.10.0
Werkzeug==0.10.4
